Make kmk_keyboard_user importable on SAMD51 by reducing recursion (read: import) depth
